+++ R/print_and_plot.R
@@ -82,7 +82,7 @@ coef_plot <- function(fitted_DAG, reverse_order = FALSE) {
   df <- tidyr::gather_(df, 'to', 'coef', colnames(fitted_DAG$coef))
   df$lower <- c(fitted_DAG$lower)
   df$upper <- c(fitted_DAG$upper)
-  df$path <- paste(df$from, df$to, sep = ' -> ')
+  df$path <- paste(df$from, df$to, sep = ' \U2192 ')
   df <- dplyr::arrange_(df,
                         ~match(df$from, colnames(fitted_DAG$coef)),
                         ~match(df$to, colnames(fitted_DAG$coef)))
@@ -98,4 +98,76 @@ coef_plot <- function(fitted_DAG, reverse_order = FALSE) {
     ggplot2::geom_pointrange(size = 0.75) +
     ggplot2::xlab('') +
     ggplot2::ylab('standardized regression coefficient \U00B1 CI')
+}
+
+
+#' Plot several causal hypothesis at once.
+#'
+#' @param models A list of \code{DAG} objects.
+#' @param algorithm A layout algorithm from \code{igraph}, see \code{\link[ggraph]{create_layout}}.
+#'   By default, uses the Kamada-Kawai layout algorithm. Another good option is \code{"sugiyama"},
+#'   which is designed to minimize overlapping edges in DAGs. However, it can often plot nodes too
+#'   close together.
+#' @param text_size Size of the node label text.
+#' @param box_x To avoid the arrows colliding with the nodes, specify the rectangular dimensions of
+#'   an invisible box around each node. If you have long labels, you need to increase this.
+#' @param box_y To avoid the arrows colliding with the nodes, specify the rectangular dimensions of
+#'   an invisible box around each node. If you have multi-line labels, you need to increase this.
+#' @param edge_width Width of the edges.
+#' @param curvature Curvature of the edges. A slight curvature can look pretty.
+#' @param arrow A \code{grid::arrow} object, specifying the shape and size of the arrowheads.
+#'
+#' @return A \code{ggplot} object.
+#' @export
+#'
+#' @examples
+#' m <- list(one = DAG(a ~ b + c + d), two = DAG(a ~ b, b ~ c, d ~ d))
+#' plot_model_set(m)
+#' plot_model_set(m, "sugiyama")
+plot_model_set <- function(models, algorithm = 'kk', text_size = 5, box_x = 12, box_y = 10,
+                           edge_width = 1, curvature = 0.05,
+                           arrow = grid::arrow(type = 'closed', 15, grid::unit(10, 'points'))) {
+  # Input checks
+  if (!is.list(models) | !all(purrr::map_lgl(models, ~inherits(., 'DAG')))) {
+    stop('models should be a list of DAG objects.')
+  }
+
+  var_names <- lapply(models, colnames)
+  if (length(models) > 1 &
+      (stats::var(lengths(models)) != 0 |
+       any(lengths(sapply(var_names[-1], setdiff, var_names[[1]])) != 0))) {
+    stop('All causal models need to include the same variables. Combined, your
+         models include the following variables:\n',
+         paste(sort(unique(unlist(var_names))), collapse = '\n'),
+         call. = FALSE)
+  }
+  # Sort all models the same
+  #models <- purrr::map(models, ~.[rownames(models[[1]]), colnames(models[[1]])])
+
+  # Build graph
+  result <- igraph::make_empty_graph() + igraph::vertices(row.names(models[[1]]))
+  for (i in seq_along(models)) {
+    m <- models[[i]]
+    ind  <- which(m == 1)
+    from <- ind %% nrow(models[[i]])
+    to   <- (ind - from) / nrow(models[[i]]) + 1
+    result <- igraph::add_edges(result, c(rbind(rownames(m)[from], colnames(m)[to])),
+                                attr = list(model = names(models)[[i]]))
+  }
+  igraph::E(result)$model <- factor(igraph::E(result)$model, names(models), names(models))
+
+  # Build plot.
+  my_lab <- function(labels) {
+    list(model = names(models))
+  }
+
+  ggraph::ggraph(result, 'igraph', algorithm = algorithm) +
+    ggraph::geom_edge_arc(curvature = curvature, arrow = arrow, edge_width = edge_width,
+                           end_cap = ggraph::rectangle(box_x, box_y, 'mm'),
+                           start_cap = ggraph::rectangle(box_x, box_y, 'mm')) +
+    ggraph::geom_node_text(ggplot2::aes_(label = ~name), size = text_size) +
+    ggraph::facet_edges(~model, labeller = my_lab) +
+    ggplot2::scale_x_continuous(expand = c(0.2, 0)) +
+    ggplot2::scale_y_continuous(expand = c(0.2, 0)) +
+    ggraph::theme_graph(foreground = 'grey80', base_family = 'sans')
 }